requirements/common.txt
alembic==1.7.3
attrs==21.2.0
click==8.0.1
coverage==6.0
Flask==2.0.1
Flask-Migrate==3.1.0
flask-pytest==0.0.5
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-resources==5.2.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
packaging==21.0
pluggy==1.0.0
protobuf==3.18.0
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
python-dotenv==0.19.0
SQLAlchemy==1.4.25
toml==0.10.2
Werkzeug==2.0.1
zipp==3.6.0
